+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/market/metricspage/MetricsPageService.kt
@@ -2,37 +2,12 @@ package io.horizontalsystems.bankwallet.modules.market.metricspage
 
 import io.horizontalsystems.bankwallet.core.subscribeIO
 import io.horizontalsystems.bankwallet.entities.DataState
-import io.horizontalsystems.bankwallet.modules.chart.AbstractChartService
-import io.horizontalsystems.bankwallet.modules.chart.ChartDataXxx
 import io.horizontalsystems.bankwallet.modules.market.MarketItem
 import io.horizontalsystems.bankwallet.modules.market.tvl.GlobalMarketRepository
 import io.horizontalsystems.bankwallet.modules.metricchart.MetricsType
-import io.horizontalsystems.chartview.ChartView
 import io.horizontalsystems.core.ICurrencyManager
-import io.horizontalsystems.core.entities.Currency
-import io.reactivex.Single
 import io.reactivex.disposables.Disposable
 import io.reactivex.subjects.BehaviorSubject
-
-
-class MetricsPageChartRepo(
-    override val currencyManager: ICurrencyManager,
-    private val metricsType: MetricsType,
-    private val globalMarketRepository: GlobalMarketRepository
-) : AbstractChartService() {
-    override val chartTypes = listOf(ChartView.ChartType.DAILY, ChartView.ChartType.WEEKLY, ChartView.ChartType.MONTHLY)
-    override val initialChartType: ChartView.ChartType = ChartView.ChartType.DAILY
-
-    override fun getItems(
-        chartType: ChartView.ChartType,
-        currency: Currency,
-    ): Single<ChartDataXxx> {
-        return globalMarketRepository.getGlobalMarketPoints(currency.code, chartType, metricsType)
-            .map {
-                ChartDataXxx(chartType, it)
-            }
-    }
-}
 
 class MetricsPageService(
     val metricsType: MetricsType,
